Added V1 of home page hero section
