+++ react-ts-app/src/components/Heros/Heros.module.scss
@@ -1,0 +1,4 @@
+@import "src/globalStyles/mainStyles.scss";
+@import "src/globalStyles/colors.scss";
+@import "src/globalStyles/fonts.scss";
+
