+++ react-ts-app/src/components/Heros/Heros.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import 'src/globalStyles/mainStyles.scss';
+
+interface HeroProps {
+  className: string;
+  title: string;
+  imagePath: string;
+}
+
+class Hero extends React.Component<HeroProps> {
+  render() {
+    return (
+      <section className={`hero-section ${this.props.className}`}>
+        {this.props.title}
+      </section>
+    );
+  }
+}
+
+export default Hero;
