+++ react-ts-app/src/globalStyles/colors.scss
@@ -10,6 +10,7 @@ $gray-900: #1f2123;
 $gray-1000: #111213;
 
 $white: #ffffff;
+$white-transparent: rgba(255, 255, 255, 0);
 
 $primary-accent-500: #9165d8;
 $primary-accent-800: #6116DC;
