+++ react-ts-app/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "compilerOptions": {
-    "baseUrl": "./",
+    "baseUrl": ".",
     "paths": {
-      "src": [ "src/" ]
+      "src": [ "src/" ],
+      "public": ["public/"]
     },
     "target": "es5",
     "lib": [
